+++ tests/integration/test_variables.py
@@ -19,8 +19,23 @@ from unittest import mock
 import pytest
 from reportportal_client import BatchedRPClient, OutputType, RPClient, ThreadedRPClient
 
-from tests import REPORT_PORTAL_SERVICE, REQUESTS_SERVICE
+from tests import REPORT_PORTAL_SERVICE
 from tests.helpers import utils
+
+
+@mock.patch(REPORT_PORTAL_SERVICE)
+def test_agent_pass_batch_payload_limit_variable(mock_client_init):
+    variables = utils.DEFAULT_VARIABLES.copy()
+    payload_size = 100
+    variables["RP_LOG_BATCH_PAYLOAD_LIMIT"] = payload_size
+    with warnings.catch_warnings(record=True) as w:
+        result = utils.run_robot_tests(["examples/simple.robot"], variables=variables)
+    assert result == 0  # the test successfully passed
+    assert len(w) == 0
+
+    payload_variable = "log_batch_payload_limit"
+    assert payload_variable in mock_client_init.call_args_list[0][1]
+    assert mock_client_init.call_args_list[0][1][payload_variable] == payload_size
 
 
 @mock.patch(REPORT_PORTAL_SERVICE)
@@ -28,32 +43,31 @@ def test_agent_pass_batch_payload_size_variable(mock_client_init):
     variables = utils.DEFAULT_VARIABLES.copy()
     payload_size = 100
     variables["RP_LOG_BATCH_PAYLOAD_SIZE"] = payload_size
-    result = utils.run_robot_tests(["examples/simple.robot"], variables=variables)
-    assert result == 0  # the test successfully passed
 
-    payload_variable = "log_batch_payload_size"
+    with pytest.warns(DeprecationWarning) as w:
+        result = utils.run_robot_tests(["examples/simple.robot"], variables=variables)
+    assert result == 0  # the test successfully passed
+    assert len(w) == 1
+    assert "Parameter `RP_LOG_BATCH_PAYLOAD_SIZE` is deprecated" in str(w[0].message)
+
+    payload_variable = "log_batch_payload_limit"
     assert payload_variable in mock_client_init.call_args_list[0][1]
     assert mock_client_init.call_args_list[0][1][payload_variable] == payload_size
 
 
-@mock.patch(REQUESTS_SERVICE)
-def test_agent_pass_launch_uuid_variable(mock_requests_init):
+@mock.patch(REPORT_PORTAL_SERVICE)
+def test_agent_pass_launch_uuid_variable(mock_client_init):
     variables = utils.DEFAULT_VARIABLES.copy()
     test_launch_id = "my_test_launch"
     variables["RP_LAUNCH_UUID"] = test_launch_id
     result = utils.run_robot_tests(["examples/simple.robot"], variables=variables)
     assert result == 0  # the test successfully passed
 
-    mock_requests = mock_requests_init.return_value
-    assert mock_requests.post.call_count == 3
-    suite_start = mock_requests.post.call_args_list[0]
-    assert suite_start[0][0].endswith("/item")
-    assert suite_start[1]["json"]["launchUuid"] == test_launch_id
+    assert "launch_uuid" in mock_client_init.call_args_list[0][1]
+    assert mock_client_init.call_args_list[0][1]["launch_uuid"] == test_launch_id
 
 
-@pytest.mark.parametrize(
-    "variable, warn_num", [("RP_PROJECT", 1), ("RP_API_KEY", 2), ("RP_ENDPOINT", 1), ("RP_LAUNCH", 1)]
-)
+@pytest.mark.parametrize("variable, warn_num", [("RP_PROJECT", 1), ("RP_ENDPOINT", 1), ("RP_LAUNCH", 1)])
 @mock.patch(REPORT_PORTAL_SERVICE)
 def test_no_required_variable_warning(mock_client_init, variable, warn_num):
     variables = utils.DEFAULT_VARIABLES.copy()
@@ -61,10 +75,10 @@ def test_no_required_variable_warning(mock_client_init, variable, warn_num):
 
     with warnings.catch_warnings(record=True) as w:
         result = utils.run_robot_tests(["examples/simple.robot"], variables=variables)
-        assert result == 0  # the test successfully passed
+    assert result == 0  # the test successfully passed
 
-        assert len(w) == warn_num
-        assert w[0].category == RuntimeWarning
+    assert len(w) == warn_num
+    assert w[0].category == RuntimeWarning
 
     mock_client = mock_client_init.return_value
     assert mock_client.start_launch.call_count == 0
@@ -92,62 +106,24 @@ def test_rp_api_key(mock_client_init):
 
     with warnings.catch_warnings(record=True) as w:
         result = utils.run_robot_tests(["examples/simple.robot"], variables=variables)
-        assert int(result) == 0, "Exit code should be 0 (no errors)"
+    assert int(result) == 0, "Exit code should be 0 (no errors)"
 
-        assert mock_client_init.call_count == 1
+    assert mock_client_init.call_count == 1
 
-        constructor_args = mock_client_init.call_args_list[0][1]
-        assert constructor_args["api_key"] == api_key
-        assert len(filter_agent_calls(w)) == 0
-
-
-@mock.patch(REPORT_PORTAL_SERVICE)
-def test_rp_uuid(mock_client_init):
-    api_key = "rp_api_key"
-    variables = dict(utils.DEFAULT_VARIABLES)
-    del variables["RP_API_KEY"]
-    variables.update({"RP_UUID": api_key}.items())
-
-    with warnings.catch_warnings(record=True) as w:
-        result = utils.run_robot_tests(["examples/simple.robot"], variables=variables)
-        assert int(result) == 0, "Exit code should be 0 (no errors)"
-
-        assert mock_client_init.call_count == 1
-
-        constructor_args = mock_client_init.call_args_list[0][1]
-        assert constructor_args["api_key"] == api_key
-        assert len(filter_agent_calls(w)) == 1
+    constructor_args = mock_client_init.call_args_list[0][1]
+    assert constructor_args["api_key"] == api_key
+    assert len(w) == 0
 
 
-@mock.patch(REPORT_PORTAL_SERVICE)
-def test_rp_api_key_priority(mock_client_init):
-    api_key = "rp_api_key"
-    variables = dict(utils.DEFAULT_VARIABLES)
-    variables.update({"RP_API_KEY": api_key, "RP_UUID": "rp_uuid"}.items())
-
-    with warnings.catch_warnings(record=True) as w:
-        result = utils.run_robot_tests(["examples/simple.robot"], variables=variables)
-        assert int(result) == 0, "Exit code should be 0 (no errors)"
-
-        assert mock_client_init.call_count == 1
-
-        constructor_args = mock_client_init.call_args_list[0][1]
-        assert constructor_args["api_key"] == api_key
-        assert len(filter_agent_calls(w)) == 0
-
-
-@mock.patch(REPORT_PORTAL_SERVICE)
-def test_rp_api_key_empty(mock_client_init):
+def test_rp_api_key_empty():
     api_key = ""
     variables = dict(utils.DEFAULT_VARIABLES)
     variables.update({"RP_API_KEY": api_key}.items())
 
-    with warnings.catch_warnings(record=True) as w:
-        result = utils.run_robot_tests(["examples/simple.robot"], variables=variables)
-        assert int(result) == 0, "Exit code should be 0 (no errors)"
-
-        assert mock_client_init.call_count == 0
-        assert len(filter_agent_calls(w)) == 2
+    with pytest.warns(Warning) as w:
+        utils.run_robot_tests(["examples/simple.robot"], variables=variables)
+    assert len(w) == 1
+    assert "Authentication credentials are required." in str(w[0].message)
 
 
 @mock.patch(REPORT_PORTAL_SERVICE)