+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## [Unreleased]
 ### Added
+- OAuth 2.0 Password Grant authentication, by @HardNorth
+### Changed
+- Client version updated on [5.6.7](https://github.com/reportportal/client-Python/releases/tag/5.6.7), by @HardNorth
+### Fixed
+- Some configuration parameter names, which are different in the client, by @HardNorth
+### Removed
+- `RP_UUID` param support, as it was deprecated pretty while ago, by @HardNorth
+
+## [5.6.4]
+### Added
 - `RP_DEBUG_MODE` configuration variable, by @HardNorth
 
 ## [5.6.3]